refactor: move `SingleColumnView` to separate dir
